==> 8. Use Cases/prediction-with-quality-inspection/1_read_data.ipynb <==
# A data quality validation use case

### **Data quality validation:** Data preparation and the importance of profiling the quality of your data

Data profiling is defined as the process of reviewing a source of data, validating it, understanding it's structure and behaviour, as well as content and depicted relations. 
Data profiling can be seen as a crucial step for different data projects and initiatives: 

- **Data Warehousing projects** - While building or feeding an existing Data Warehouse(DWH);
- **Data conversion and projects for data migration** - Data profiling is an essential tool to spot and identify any data quality issues. It can also help uncivering any new requirements for the target system;
- **Data preparation** - To develop ML based projects the data preparation stage involves many activites around data filtering and delection, data augmentation, feature creation and data splits based on training, validation and holdout. To measure the impact of each of these transformations data-profiling is a core step that allows to automate any required validations. 

YData's data profiling process involves: 
- Inference of data types and keyword tagging based categories;
- The collection of descriptive statistics in an univariate manner (min, max, quantiles, etc);
- Key integraty and missingess profile, based on zeros, blanks and nulls validation;
- Collection of statistics on a multivariate manner (how each variable relates with each other);
- Constraints defined based on business rules and formatting expectations.

## Hotel booking - A data preparation & profiling pipeline

To demo the potential of data quality profiling analysis throuhgout the process of data preparation, the example below leverages the ["Hotel Booking dataset that can be found in Kaggle"](https://www.kaggle.com/datasets/jessemostipak/hotel-booking-demand). The dataset contains the booking information for a city and resort hotel.
In this tutorial we deliver a full pipeline with the data preparation to train a classification model to identify wether a certain reservation will be canceled. The pipeline includes the validation of the train/validation split (is the samples representative and comaprable?) as well as the profiling of all the data transformations and features extracted throughout the process. 

## Read data

# Import the necessary packages
from pickle import dump

from ydata.platform.datasources import DataSources
from ydata.metadata import Metadata

# Creating a Dataset from the Data Source
datasource = DataSources.get(uid='8a730a1d-d756-4768-a79d-663c86fa2af0', namespace='c69936a5-a14e-43aa-bcc7-f7e95709bb9b')
dataset = datasource.read()
# Quickly previewing the Dataset
dataset.head()

## Metadata calculation
Calculating the overall statistics of the data to be observed

# Creating a Metadata (where warnings can be reviewed) from the Dataset
metadata = Metadata(dataset)
print(metadata)

### Update metadata

dataset.astype('reservation_status_date', 'datetime')

metadata = Metadata(dataset)

#aqui fazer o update da metadata
metadata.update_datatypes({'is_canceled': 'categorical'})

## Pipeline outputs

#Savin the dataset metadata
metadata.save('bookings_metadata.pkl')

from pickle import dumps

with open('bookings.pkl', 'wb') as file:
    dump(dataset, file)
