==> 8. Use Cases/prediction-with-quality-inspection/3_traindata_preparation.ipynb <==
# 3. Data preparation

### Import the needed packages

import os
import json
from pickle import load, dump

import numpy as np
import pandas as pd
from pandas import DataFrame
from ydata.metadata import Metadata

from functions.get_dummies import GetDummies

from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import StandardScaler
from sklearn.impute import SimpleImputer
from sklearn.feature_selection import VarianceThreshold, SelectKBest, chi2

### Getting the environment variables values

data_split = os.environ.get('DATA_SPLIT', 'train')
dups = os.environ.get('KEEP_DUP', 'first')

### Auxiliary functions definition

#Function to load the training set
def load_sets(path: str):
    with open(path, 'rb') as f:
        return load(f)
    
def pickle_sets(df: pd.DataFrame, y: pd.Series, path:str):
    dataset = (df, y)
    
    with open(path, 'wb') as f:
        dump(dataset, f)

def profile_json(df: DataFrame):
    profile = ProfileReport(df = df)
    profile_json = json.loads(profile.json)
    
    return profile_json

def calculate_profile_diff(profiles: list):
    return profiles

def unique_levels(x):
    x = x.value_counts().count()
    return x

## Read the dataset, metadata & profiling

metadata = Metadata.load('bookings_metadata.pkl')

data, y = load_sets(f'{data_split}.pkl')
data['is_canceled']=y

## Prepare the data

### Check & act upon the metadata warnings

#remove duplicates
for warning_type, warnings in metadata.warnings.items():
    if warning_type == 'duplicates' and len(warnings)>0:
        ##filter the dataset duplicates
        data.drop_duplicates(keep=dups, inplace=True)

y = data['is_canceled']
data.drop(['is_canceled', 'booking_changes'], axis=1, inplace=True)

fltr_metadata=metadata[[col for col in metadata.columns if col not in ['is_canceled', 'booking_changes']]]

## Split Features into numerical and categorical
num = data[fltr_metadata.numerical_vars]
char = data[fltr_metadata.categorical_vars]

#Check num and categorical
# check for how many unique values each column has
df_value_counts = pd.DataFrame(num.apply(lambda X: unique_levels(X)))

df_value_counts.columns = ['feature_levels']
df_value_counts.sort_values(by = 'feature_levels', ascending=False)

slice1 = df_value_counts.loc[df_value_counts['feature_levels']<=20]
cat_list = slice1.index
cat = num.loc[:, cat_list]

slice2 = df_value_counts.loc[df_value_counts['feature_levels']>20]
num_list = slice2.index
num = num.loc[:, num_list]

char = pd.concat([char, cat], axis = 1, join = 'inner')

char = char.replace(np.nan, '')

## Dropping Variables that have >25% missing data
#only numerical has missing data
num = num.loc[:, num.isnull().mean() <= 0.25]

cols_split = {'cat': list(char.columns), 'num': list(num.columns)}

##Defining the imputers for the missing data identified
imputers= {'cat': SimpleImputer(missing_values='', strategy='most_frequent'),
           'num': SimpleImputer(missing_values=np.nan, strategy='mean')}

for k, imputer in imputers.items():
    if k=='cat':
        char_i = pd.DataFrame(imputer.fit_transform(char), index = char.index, columns=char.columns)
    elif k=='num':
        num_i = pd.DataFrame(imputer.fit_transform(num), index = num.index, columns=num.columns)

### Feature Selection

## Remove variables with zero variance
varselector = VarianceThreshold(threshold = 0)
varselector.fit_transform(num_i)

cols = varselector.get_support(indices=True)

##select_cols
num_i = num_i.iloc[:, cols]

##Create new features
num_i['day_wait_ind'] = np.where(num_i['days_in_waiting_list']>0, 1, 0)
num_i['previous_bookings_not_canceled_ind'] = np.where(num_i['previous_bookings_not_canceled']>0, 1, 0)
                                                             
try:
    num_i['booking_changes_ind'] = np.where(num_i['booking_changes']<0, 1, 0)

    num_varlist = ['arrival_date_day_of_month', 'arrival_date_week_number', 'days_in_waiting_list', 'previous_bookings_not_canceled', 'booking_changes']
except:
    num_varlist = ['arrival_date_day_of_month', 'arrival_date_week_number', 'days_in_waiting_list', 'previous_bookings_not_canceled']

num_i.drop(num_varlist, axis=1, inplace=True)

char_i

#Categorical features
try:
    charlist = ['arrival_date_month', 'country', 'assigned_room_type', 'reservation_status',
                'reservation_status_date', 'arrival_date_year']
    char_i.drop(charlist, axis=1, inplace=True)
except:
    charlist = ['arrival_date_month', 'country', 'assigned_room_type', 'reservation_status', 'arrival_date_year']
    char_i.drop(charlist, axis=1, inplace=True)

#LabelEncoder
#Encode all the categorical columns
le = GetDummies()

le.fit(char_i)
char_encoded = le.transform(char_i)

selector = SelectKBest(chi2, k=30)
selector.fit_transform(char_encoded, y)

#get columns to create new df with them only
cols = selector.get_support(indices=True)
select_features_df_char = char_encoded.iloc[:,cols]

##Setting the final DF
X_train = pd.concat([select_features_df_char, num_i], axis=1, join='inner')

selected_vars= {'cat': list(select_features_df_char.columns), 'num': list(num_i.columns)}

### Output methods for data preparation

with open('cols_split.pkl', 'wb') as f:
    dump(cols_split, f)

with open('imputers.pkl', 'wb') as f:
    dump(imputers, f)
    
with open('encoders.pkl', 'wb') as f:
    dump(le, f)
    
with open('selected_vars.pkl', 'wb') as f:
    dump(selected_vars, f)

## Output data

## Add here more details
pickle_sets(df=X_train, y=y, path=f'dataprep_{data_split}.pkl')
