==> 8. Use Cases/prediction-with-quality-inspection/4_data_profiling.ipynb <==
# 4. Exploratory data analysis - First understanding of the data

### Import the needed packages

import os

from pickle import load
import pandas as pd

from ydata.metadata import Metadata
from ydata.profiling import ProfileReport

from functions.save_load import load_sets

### Getting the environment variables values

set_path = os.environ.get('SET_PATH', 'train.pkl')
data_split=os.environ.get('DATA_SPLIT', 'train')

## Load the dataset & metadata

data, y = load_sets(set_path)

try:
    data = data.drop('Unnamed: 0', axis=1)
except:
    "No unnamed"

meta = Metadata.load('bookings_metadata.pkl')
print(meta)

## Generating the full data profile

print(f'Profile Name: {data_split}_profile')
profile = ProfileReport(df=data, title='Hotel bookings demand')
profile.config.html.navbar_show = False

profile.to_file(f'{data_split}_profile.html')

### Plotting the target ratio

ratio_labels = pd.DataFrame(y.value_counts(normalize=True))

ratio_labels

## Pipeline output

import json

metadata = {
    'outputs' : [
        {
      'type': 'table',
      'storage': 'inline',
      'format': 'csv',
      'header': list(ratio_labels.columns),
      'source': ratio_labels.to_csv(header=False, index=True)
    },
        {
      'type': 'web-app',
      'storage': 'inline',
      'source': profile.to_html(),
    }
    ]
  }

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(metadata, metadata_file)
